+++ catalyst/contrib/utils/dataset.py
@@ -123,3 +123,6 @@ def create_dataframe(dataset: DictDataset, **dataframe_args) -> pd.DataFrame:
     ]
     df = pd.DataFrame(data, **dataframe_args)
     return df
+
+
+__all__ = ["create_dataset", "create_dataframe", "split_dataset_train_test"]
