+++ catalyst/contrib/utils/pandas.py
@@ -572,3 +572,21 @@ def balance_classes(
     balanced_df = pd.concat(balanced_dfs.values())
 
     return balanced_df
+
+
+__all__ = [
+    "dataframe_to_list",
+    "folds_to_list",
+    "split_dataframe",
+    "split_dataframe_on_column_folds",
+    "split_dataframe_on_folds",
+    "split_dataframe_on_stratified_folds",
+    "split_dataframe_train_test",
+    "separate_tags",
+    "read_multiple_dataframes",
+    "map_dataframe",
+    "get_dataset_labeling",
+    "merge_multiple_fold_csv",
+    "read_csv_data",
+    "balance_classes",
+]
