+++ catalyst/contrib/utils/text.py
@@ -92,3 +92,6 @@ def process_bert_output(
             output[i] = _process_features(feature_)
 
     return output
+
+
+__all__ = ["tokenize_text", "process_bert_output"]
