+++ catalyst/core/callbacks/exception.py
@@ -1,5 +1,4 @@
-from catalyst import utils
-from catalyst.core import Callback, CallbackNode, CallbackOrder, State
+from catalyst.core import Callback, CallbackNode, CallbackOrder, State, utils
 
 
 class ExceptionCallback(Callback):
