+++ catalyst/core/callbacks/logging.py
@@ -5,9 +5,8 @@ import sys
 
 from tqdm import tqdm
 
-from catalyst import utils
-from catalyst.core import Callback, CallbackNode, CallbackOrder, State
-from catalyst.utils.tools.tensorboard import SummaryWriter
+from catalyst.contrib.utils.tools.tensorboard import SummaryWriter
+from catalyst.core import Callback, CallbackNode, CallbackOrder, State, utils
 
 from . import formatters
