+++ catalyst/core/utils/__init__.py
@@ -1,5 +1,6 @@
 # flake8: noqa
 
+from catalyst.contrib.utils import *
 from catalyst.utils import *
 
 from .callbacks import filter_callbacks_by_node, sort_callbacks_by_order
