+++ catalyst/utils/hash.py
@@ -41,3 +41,6 @@ def get_short_hash(o) -> str:
     """
     hash = get_hash(o)[:6]
     return hash
+
+
+__all__ = ["get_hash", "get_short_hash"]
