+++ catalyst/utils/misc.py
@@ -150,3 +150,15 @@ def fn_ends_with_pass(fn: Callable[..., Any]):
     if source_lines[-1].strip() == "pass":
         return True
     return False
+
+
+__all__ = [
+    "copy_directory",
+    "format_metric",
+    "get_fn_default_params",
+    "get_fn_argsnames",
+    "get_utcnow_time",
+    "is_exception",
+    "maybe_recursive_call",
+    "fn_ends_with_pass",
+]
