+++ catalyst/utils/tools/__init__.py
@@ -1,11 +1,4 @@
 # flake8: noqa
 from .frozen_class import FrozenClass
 from .registry import Registry, RegistryException
-from .tensorboard import (
-    EventReadingException,
-    EventsFileReader,
-    SummaryItem,
-    SummaryReader,
-    SummaryWriter,
-)
 from .time_manager import TimeManager
